tests/wpt/web-platform-tests/webrtc/RTCPeerConnection-setLocalDescription.html

<!doctype html>
<meta charset=utf-8>
<title>RTCPeerConnection.prototype.setLocalDescription</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="RTCPeerConnection-helper.js"></script>
<script>
  'use strict';

  // Test is based on the following editor draft:
  // https://w3c.github.io/webrtc-pc/archives/20170605/webrtc.html

  // The following helper functions are called from RTCPeerConnection-helper.js:
  //   generateOffer
  //   assert_session_desc_not_similar
  //   assert_session_desc_similar

  /*
    4.3.2.  Interface Definition
      [Constructor(optional RTCConfiguration configuration)]
      interface RTCPeerConnection : EventTarget {
        Promise<void>                      setRemoteDescription(
            RTCSessionDescriptionInit description);

        readonly attribute RTCSessionDescription? remoteDescription;
        readonly attribute RTCSessionDescription? currentRemoteDescription;
        readonly attribute RTCSessionDescription? pendingRemoteDescription;
        ...
      };

    4.6.2.  RTCSessionDescription Class
      dictionary RTCSessionDescriptionInit {
        required RTCSdpType type;
                 DOMString  sdp = "";
      };

    4.6.1.  RTCSdpType
      enum RTCSdpType {
        "offer",
        "pranswer",
        "answer",
        "rollback"
      };
   */

  promise_test(t => {
    const pc = new RTCPeerConnection();
    t.add_cleanup(() => pc.close());

    const states = [];
    pc.addEventListener('signalingstatechange', () => states.push(pc.signalingState));

    return pc.createOffer({ offerToReceiveAudio: true })
    .then(offer1 =>
      pc.setLocalDescription(offer1)
      .then(() => generateAnswer(offer1))
      .then(answer => pc.setRemoteDescription(answer))
      .then(() => {
        pc.createDataChannel('test');
        return pc.createOffer({ offerToReceiveVideo: true });
      })
      .then(offer2 =>
        pc.setLocalDescription(offer2)
        .then(() => {
          assert_equals(pc.signalingState, 'have-local-offer');
          assert_session_desc_not_similar(offer1, offer2);
          assert_session_desc_similar(pc.localDescription, offer2);
          assert_session_desc_similar(pc.currentLocalDescription, offer1);
          assert_session_desc_similar(pc.pendingLocalDescription, offer2);

          assert_array_equals(states, ['have-local-offer', 'stable', 'have-local-offer']);
        })));
  }, 'Calling createOffer() and setLocalDescription() again after one round of local-offer/remote-answer should succeed');

  promise_test(t => {
    const pc = new RTCPeerConnection();
    t.add_cleanup(() => pc.close());

    const states = [];
    pc.addEventListener('signalingstatechange', () => states.push(pc.signalingState));

    return generateOffer({ pc, data: true })
    .then(offer => pc.setRemoteDescription(offer))
    .then(() => pc.createAnswer())
    .then(answer =>
      pc.setLocalDescription(answer)
      .then(() => pc.createOffer({ offerToReceiveVideo: true }))
      .then(offer =>
        pc.setLocalDescription(offer)
        .then(() => {
          assert_equals(pc.signalingState, 'have-local-offer');
          assert_session_desc_similar(pc.localDescription, offer);
          assert_session_desc_similar(pc.currentLocalDescription, answer);
          assert_session_desc_similar(pc.pendingLocalDescription, offer);

          assert_array_equals(states, ['have-remote-offer', 'stable', 'have-local-offer']);
        })));
  }, 'Switching role from answerer to offerer after going back to stable state should succeed');

  promise_test(async t => {
    const pc = new RTCPeerConnection();
    const offer = await pc.createOffer();
    let eventSequence = '';
    const signalingstatechangeResolver = new Resolver();
    pc.onsignalingstatechange = () => {
      eventSequence += 'onsignalingstatechange;';
      signalingstatechangeResolver.resolve();
    };
    await pc.setLocalDescription(offer);
    eventSequence += 'setLocalDescription;';
    await signalingstatechangeResolver.promise;
    assert_equals(eventSequence, 'onsignalingstatechange;setLocalDescription;');
  }, 'onsignalingstatechange fires before setLocalDescription resolves');

  /*
    TODO
      4.3.2.  setLocalDescription
        4.  If description.sdp is null and description.type is pranswer, set description.sdp
            to lastAnswer.
        7.  If description.type is pranswer and description.sdp does not match lastAnswer,
            reject the promise with a newly created InvalidModificationError and abort these
            steps.
   */

</script>
